refactor(sysAdmin): mdコピーボタンを縮小しテンプレリンクの右に併設
- mdコピーボタンを CopyMarkdownButton として切り出し、ラベルを
  「mdコピー」→「md」に短縮して h-6 / icon-3 のコンパクトサイズへ縮小
- ReportContentSection から取り外し、ヘッダー (ReportDetailHeader) の
  テンプレリンク行に同居
- テンプレリンクの文言を「このレポートのテンプレートを開く (v3)」→
  「テンプレ(v3)」に短縮
