+++ src/app/sysAdmin/features/reportDetail/components/ReportDetailHeader.tsx
@@ -11,6 +11,7 @@ import {
   variantLabel,
 } from "@/app/sysAdmin/features/system/templates/shared/labels";
 import { variantToSlug } from "@/app/sysAdmin/features/system/templates/shared/variantSlug";
+import { CopyMarkdownButton } from "./CopyMarkdownButton";
 
 type Props = {
   variant: GqlReportVariant;
@@ -18,12 +19,14 @@ type Props = {
   periodFrom: Date;
   periodTo: Date;
   templateVersion?: number | null;
+  /** あれば「md コピー」アクションを並べる。null/空なら出さない。 */
+  body?: string | null;
 };
 
 /**
  * Report detail page の header。
  * 期間 / variant / status を inline テキストで出し、その下にこのレポートを生成した
- * テンプレ詳細へのリンクを置く (templateVersion が分かれば付ける)。
+ * テンプレ詳細へのリンクと、md コピーボタンを並べる。
  */
 export function ReportDetailHeader({
   variant,
@@ -31,6 +34,7 @@ export function ReportDetailHeader({
   periodFrom,
   periodTo,
   templateVersion,
+  body,
 }: Props) {
   const slug = variantToSlug(variant);
   const templateHref = slug
@@ -52,17 +56,22 @@ export function ReportDetailHeader({
         </span>
         <span>{statusLabel(status)}</span>
       </p>
-      {templateHref && (
-        <Link
-          href={templateHref}
-          className="inline-flex items-center gap-1 text-body-xs text-muted-foreground hover:text-foreground"
-        >
-          <span>
-            このレポートのテンプレートを開く
-            {templateVersion != null ? ` (v${templateVersion})` : ""}
-          </span>
-          <ExternalLink className="h-3 w-3" aria-hidden />
-        </Link>
+      {(templateHref || body) && (
+        <div className="flex items-center gap-2">
+          {templateHref && (
+            <Link
+              href={templateHref}
+              className="inline-flex items-center gap-1 text-body-xs text-muted-foreground hover:text-foreground"
+            >
+              <span>
+                テンプレ
+                {templateVersion != null ? `(v${templateVersion})` : ""}
+              </span>
+              <ExternalLink className="h-3 w-3" aria-hidden />
+            </Link>
+          )}
+          {body && <CopyMarkdownButton text={body} />}
+        </div>
       )}
     </header>
   );
